replace viewport units only in the editor
