+++ src/extensions/maxi-block/maxiBlockComponent.js
@@ -25,6 +25,7 @@ import {
 	getRelations,
 	styleGenerator,
 	styleResolver,
+	viewportUnitsProcessor,
 } from '../styles';
 import getBreakpoints from '../styles/helpers/getBreakpoints';
 import getIsUniqueIDRepeated from './getIsUniqueIDRepeated';
@@ -80,7 +81,11 @@ const StyleComponent = ({
 
 	const styles = styleResolver(stylesObj, false, getBreakpoints());
 
-	const styleContent = styleGenerator(styles, isIframe, isSiteEditor);
+	const styleContent = styleGenerator(
+		viewportUnitsProcessor(styles), // replacing viewport units only for the editor
+		isIframe,
+		isSiteEditor
+	);
 
 	return <style>{styleContent}</style>;
 };
