+++ src/extensions/styles/index.js
@@ -28,5 +28,6 @@ export { default as transitionAttributesCreator } from './transitions/transition
 export * from './styleGenerator';
 export { default as styleResolver } from './styleResolver';
 export { default as styleProcessor } from './styleProcessor';
+export { default as viewportUnitsProcessor } from './viewportUnitsProcessor';
 
 export * from './utils';
